+++ lib/src/git_data.dart
@@ -17,7 +17,7 @@ abstract class GitPerson {
         mail = getPersonMail(str);
 
   static String getPersonName(String str) {
-    if (-1 == str.indexOf("<")) return "Unknown";
+    if (!str.contains("<")) return "Unknown";
     return str.split("<")[0].trim();
   }
 
@@ -119,9 +119,9 @@ class GitRemote {
 }
 
 class GitBranch {
-  String name;
-  String reference;
-  String id;
+  final String name;
+  final String reference;
+  final String id;
 
   GitBranch(this.name, this.reference, this.id);